+++ CHANGELOG.md
@@ -10,12 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - [client] Run morio init on install
-- [core] Build repo installer package on initial setup
+- [drbuilder] Added new drbuilder service
+- [core] Build both client and repo installer package on initial setup
 
 ### Fixed
 
 - [client] Do not attempt to enable services at install
 - [client] Handle status subcommand when no arguments are passed
+- [dbuilder] Fixed issue in the client package build step that resolved in an invalid APT package
 - [dbuilder] Fix issue with the repo installer package build
 - [moriod] Detect non-interactive invovation in install script
 - [ui] Downloads page now lists the Repository Installer packages